requirement.txt
Flask-2.2.2 
Jinja2-3.1.2 
MarkupSafe-2.1.2 
Werkzeug-2.2.2 
click-8.1.3 
colorama-0.4.6 
itsdangerous-2.1.2
sklearn-0.0.post1
tensorflow-intel==2.11.0
pydantic-1.10.4
flask_pydantic-0.11.0
pandas-1.5.3
vader
textblob